pyproject.toml
[project]
name = "extreme-motion-reimpl"
version = "0.1.0"
description = "Overnight ladder for paper-faithful, breakdance-oriented vision reimplementation experiments"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.2.0",
    "PyYAML>=6.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.0",
]

[project.scripts]
extreme-motion-reimpl = "extreme_motion_reimpl.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
